Added multiple enemies and made different images

+++ src/entity/model/Entity.java
@@ -68,7 +68,7 @@ public abstract class Entity {
         return this.speed;
     }
     
-    public void setSpeed(int speed) {
+    public void setSpeed(double speed) {
         this.speed = speed;
     }
